Apply the following patch from Intel to improve LACP performance.
From 7dbe41273086c698f565843d5790034786f01719 Mon Sep 17 00:00:00 2001
From: Keith Wiles <keith.wiles@intel.com>
Date: Fri, 6 Jan 2017 12:20:41 -0600
Subject: [PATCH] driver/bond:lacp performance fix

When running in a bonded mode (LACP) with two ports in the bond. When
one of the ports link is down the performance increases. If the port
is returned to the bond in link up mode the performance decreases in
packet rate.

The problem appears to be when polling the slave ports the rx routine
would always start with the first slave port, which could then starve
the others ports as the number of packets requested is mostly filled
by the first port.

The fix here round robins the slave ports on each entry into the bond
driver rx routine called from the application. Giving the other ports
a chance to be serviced before packets are dropped.

The patch was tested with RX descriptors set to 256 and not 128 as was
the default value.

Signed-off-by: Keith Wiles <keith.wiles@intel.com>