DT-NIMS hotfix: histograms and logging.
The changes in this commit will increase visibility for LACP packet tx and rx on
both the state machine and the bond pmd. Additionally, the changes will enable
monitoring for any anomalies in the behaviour of the state machine, bond pmd,
eal alarms and the eal interrupt thread. It will also help us measure the time
taken by the glibc calloc() function call which is a backported fix from later
versions of dpdk to address the rte_zmalloc() fuction call(replaced with calloc()
in the backported fix) delays observed in the DT setups which was one of the causes
for frequent LACP jitters leading to LACP interface flaps.

Note: This patch is a hot-fix to address the DT issue CEM-21858 and the patch will
be reverted when a build becomes available after the merge. The patch is being
reverted since this monitoring tool is specifially designed for the DT environment
and has some shortcomings such as it is only valid for a setup functioning in fast
LACP mode and it is not production ready for all contrail environments. A JIRA story
https://contrail-jws.atlassian.net/browse/CN2-2822 has been created to address this
issue and make this tool a full fledged monitoring framework for all contrail
environments for future releases.

Closes-Jira-Bug: CEM-21858
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: Ic042d839eb44bb6334dc92f5e8e237253feefa1a