WIP: Correction to set alarm only once for multiple slaves.
During LACP Timer thread debugging, it was noticed that the rte_eal_alarm_set()
function is being called for each slave when the bond_mode_8023ad_start() function
is called during dpdk start up. Due to this, multiple instances of periodic CB are
triggered every time the eal-intr-thread is called for Timer thread.
Since the bond_mode_8023ad_periodic_cb() function handles rx and tx for each slave on
the timer thread, only one instance of this function should be invoked periodically.

To fix this, we can call the rte_eal_alarm_cancel() to clear any alarms which are
set during the slave additions for which the bond_mode_8023ad_start() was called
previously.

Closes-Jira-Bug: CEM-23915
Signed-off-by: Vinay Agrawal <vagrawal@juniper.net>
Change-Id: If5e6396a1bd0edc6f5d4edce9c3cc70b11ac090b